fix all win64 warnings

+++ Source/AliveLibAE/Psx.cpp
@@ -907,7 +907,7 @@ namespace Test
         PrimHeader* ot[5] = {};
         PSX_ClearOTag_4F6290(ot, 5);
         ASSERT_EQ(ot[0], (PrimHeader*)&ot[1]);
-        ASSERT_EQ(ot[4], (PrimHeader*)0xFFFFFFFF);
+        ASSERT_EQ(ot[4], (PrimHeader*)static_cast<size_t>(0xFFFFFFFF));
     }
 
     void PsxTests()
